requirements.txt
Flask==2.1.1
flask-table==0.5.0
pytesseract==0.3.8
Pillow==8.3.2